+++ hipcub/include/hipcub/backend/rocprim/thread/thread_reduce.hpp
@@ -1,7 +1,7 @@
 /******************************************************************************
  * Copyright (c) 2010-2011, Duane Merrill.  All rights reserved.
  * Copyright (c) 2011-2018, NVIDIA CORPORATION.  All rights reserved.
- * Modifications Copyright (c) 2017-2024, Advanced Micro Devices, Inc.  All rights reserved.
+ * Modifications Copyright (c) 2017-2025, Advanced Micro Devices, Inc.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -34,55 +34,159 @@
 
 BEGIN_HIPCUB_NAMESPACE
 
-/// Internal namespace (to prevent ADL mishaps between static functions when mixing different CUB installations)
-namespace internal {
-
-template <
-    int         LENGTH,
-    typename    T,
-    typename    ReductionOp,
-    bool        NoPrefix = false>
-__device__ __forceinline__ T ThreadReduce(
-    T*           input,
-    ReductionOp reduction_op,
-    T           prefix = T(0))
+namespace internal
 {
-    T retval;
-    if(NoPrefix)
-        retval = input[0];
-    else
-        retval = prefix;
 
-    #pragma unroll
-    for (int i = 0 + NoPrefix; i < LENGTH; ++i)
+template<typename AccumType, typename InputType, typename ReductionOp, typename PrefixType>
+[[nodiscard]]
+__device__ __forceinline__
+AccumType
+    ThreadReduceSequential(const InputType& input, ReductionOp reduction_op, PrefixType prefix)
+{
+    AccumType     retval = static_cast<AccumType>(prefix);
+    constexpr int length = ::hipcub::detail::static_size_v<InputType>();
+#pragma unroll
+    for(int i = 0; i < length; ++i)
+    {
         retval = reduction_op(retval, input[i]);
-
+    }
     return retval;
 }
 
-template <
-    int         LENGTH,
-    typename    T,
-    typename    ReductionOp>
-__device__ __forceinline__ T ThreadReduce(
-    T           (&input)[LENGTH],
-    ReductionOp reduction_op,
-    T           prefix)
+template<typename AccumType, typename InputType, typename ReductionOp>
+[[nodiscard]]
+__device__ __forceinline__
+AccumType ThreadReduceSequential(const InputType& input, ReductionOp reduction_op)
 {
-    return ThreadReduce<LENGTH, false>((T*)input, reduction_op, prefix);
+    AccumType     retval = input[0];
+    constexpr int length = ::hipcub::detail::static_size_v<InputType>();
+#pragma unroll
+    for(int i = 1; i < length; ++i)
+    {
+        retval = reduction_op(retval, input[i]);
+    }
+    return retval;
 }
 
-template <
-    int         LENGTH,
-    typename    T,
-    typename    ReductionOp>
-__device__ __forceinline__ T ThreadReduce(
-    T           (&input)[LENGTH],
-    ReductionOp reduction_op)
+// This function is not used, it is just here for compatibility
+template<typename AccumType, typename InputType, typename ReductionOp>
+[[nodiscard]] [[deprecated("This function is an internal API, using it directly is not "
+                           "recommended.")]]
+__device__ __forceinline__
+AccumType ThreadReduceBinaryTree(const InputType& input, ReductionOp reduction_op)
 {
-    return ThreadReduce<LENGTH, true>((T*)input, reduction_op);
+    constexpr auto length = ::hipcub::detail::static_size_v<InputType>();
+#pragma unroll
+    for(int i = 1; i < length; i *= 2)
+    {
+#pragma unroll
+        for(int j = 0; j + i < length; j += i * 2)
+        {
+            input[j] = reduction_op(input[j], input[j + i]);
+        }
+    }
+    return input[0];
 }
 
+template<typename AccumType, typename InputType, typename ReductionOp>
+[[nodiscard]] [[deprecated("This function is an internal API, using it directly is not "
+                           "recommended.")]]
+__device__ __forceinline__
+AccumType ThreadReduceTernaryTree(const InputType& input, ReductionOp reduction_op)
+{
+    constexpr auto length = ::hipcub::detail::static_size_v<InputType>();
+#pragma unroll
+    for(int i = 1; i < length; i *= 3)
+    {
+#pragma unroll
+        for(int j = 0; j + i < length; j += i * 3)
+        {
+            auto value = reduction_op(input[j], input[j + i]);
+            input[j]   = (j + i * 2 < length) ? reduction_op(value, input[j + i * 2]) : value;
+        }
+    }
+    return input[0];
+}
+
+// TODO: we should also implement ThreadReduceSimd after simd intrinsics are available.
+
+} // namespace internal
+
+template<typename InputType,
+         typename ReductionOp,
+         typename ValueType = typename ::std::remove_cv<typename ::std::remove_reference<
+             decltype(::std::declval<InputType>()[0])>::type>::type,
+         typename AccumType = ::rocprim::accumulator_t<ReductionOp, ValueType>>
+[[nodiscard]]
+__device__ __forceinline__
+AccumType ThreadReduce(const InputType& input, ReductionOp reduction_op)
+{
+    static_assert(::hipcub::detail::is_fixed_size_random_access_range<InputType>::value,
+                  "InputType must support the subscript operator[] and have a compile-time size");
+    static_assert(std::is_invocable<ReductionOp, ValueType, ValueType>::value,
+                  "ReductionOp must be invocable with operator()(ValueType, ValueType)");
+    return ::hipcub::internal::ThreadReduceSequential<AccumType>(input, reduction_op);
+}
+
+template<typename InputType,
+         typename ReductionOp,
+         typename PrefixType,
+         typename ValueType = typename ::std::remove_cv<typename ::std::remove_reference<
+             decltype(::std::declval<InputType>()[0])>::type>::type,
+         typename AccumType = ::rocprim::accumulator_t<ReductionOp, ValueType, PrefixType>>
+[[nodiscard]]
+__device__ __forceinline__
+AccumType ThreadReduce(const InputType& input, ReductionOp reduction_op, PrefixType prefix)
+{
+    static_assert(::hipcub::detail::is_fixed_size_random_access_range<InputType>::value,
+                  "InputType must support the subscript operator[] and have a compile-time size");
+    static_assert(std::is_invocable<ReductionOp, ValueType, ValueType>::value,
+                  "ReductionOp must be invocable with operator()(ValueType, ValueType)");
+    constexpr auto length = ::hipcub::detail::static_size_v<InputType>();
+    return ::hipcub::internal::ThreadReduceSequential<AccumType>(input, reduction_op, prefix);
+}
+
+template<int Length,
+         typename T,
+         typename ReductionOp,
+         typename AccumType = ::rocprim::accumulator_t<ReductionOp, T>>
+[[nodiscard]]
+__device__ __forceinline__
+AccumType ThreadReduce(const T* input, ReductionOp reduction_op)
+{
+    static_assert(Length > 0, "Length must be greater than 0");
+    static_assert(std::is_invocable<ReductionOp, T, T>::value,
+                  "ReductionOp must have the binary call operator: operator(V1, V2)");
+    auto array = reinterpret_cast<const T(*)[Length]>(input);
+    return ::hipcub::ThreadReduce(*array, reduction_op);
+}
+
+template<int Length,
+         typename T,
+         typename ReductionOp,
+         typename PrefixType,
+         typename AccumType = ::rocprim::accumulator_t<ReductionOp, T, PrefixType>,
+         typename std::enable_if<(Length > 0), int>::type = 0>
+[[nodiscard]]
+__device__ __forceinline__
+AccumType ThreadReduce(const T* input, ReductionOp reduction_op, PrefixType prefix)
+{
+    static_assert(std::is_invocable<ReductionOp, T, T>::value,
+                  "ReductionOp must have the binary call operator: operator(V1, V2)");
+    auto array = reinterpret_cast<const T(*)[Length]>(input);
+    return ::hipcub::ThreadReduce(*array, reduction_op, prefix);
+}
+
+template<int Length,
+         typename T,
+         typename ReductionOp,
+         typename PrefixType,
+         typename std::enable_if<(Length == 0), int>::type = 0>
+[[nodiscard]]
+__device__ __forceinline__
+T ThreadReduce(const T*, ReductionOp, PrefixType prefix)
+{
+    return prefix;
 }
 
 END_HIPCUB_NAMESPACE